All TOD tests are now passing in simulation
It is not possible to map out all cycle accurate behavior of the TOD in
an unmodified C64, since the TOD input is fed from an external signal.
Hence, emulators do not currently sport cycle accurate TOD behavior.

The following flags can be provided to the simulation in order to match
the behavior of emulators:

* TOD_PHI20_NODELAY: Use PHI2 instead of PHI2/4 for internal clocking
* TOD_INT_NODELAY: Don't delay the ICR ALRM flag by one cycle

These flags can be passed to make via the variable SIM_OPTS, e.g.:

  make sim SIM_OPTS="-DTOD_PHI20_NODELAY -DTOD_INT_NODELAY"

Note also that a TOD rising edge does only take effect after 4 cycles;
this can be compensated for in emulators by loggin the TOD rising edge
4 cycles before clocking the emulated TOD block.